packages/features/sdd_integration/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sdd-integration"
version = "3.0.0"
requires-python = ">=3.10"

dependencies = [
    "PyYAML",
    "sdd-core",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]